[mhv-33614] Refactoring delete function
